+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.2] - Register channel plugin. 
+
++ Registry plugin system
+
 ## [0.0.1] - Initial release. 
 
 * Upload files.